pyproject.toml
[project]
name = "langchain-langgraph-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "langchain-core>=1.2.1",
    "langchain-google-vertexai>=3.2.0",
    "langchain-openai>=1.1.3",
    "langchain[anthropic]>=1.2.0",
    "pre-commit>=4.5.0",
    "python-dotenv>=1.2.1",
]

[tool.black]
line-length = 88             # 1行の文字数（デフォルトは88。チームによっては100や120にする場合も）
target-version = ['py311']   # ターゲットとするPythonのバージョン
include = '\.pyi?$'          # 対象ファイル（基本デフォルトでOK）
exclude = '''                # 除外するディレクトリ
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
